requirements.txt
flask~=3.0.0
flask_cors~=4.0.0
python-dotenv~=1.0.0
objaverse~=0.1.5
trimesh~=4.0.0
numpy~=1.26.1